Move bootstrap script inside the module
This, despite the point of the script being to be used independently of
the remainder of the module -- to bootstrap transmute and applications
built with it.

This is done to allow the code therein to be imported and reused to make
PyPI more readily available as a source for packages. Other than code
reuse, the script keeps its purpose and must remain independent.