Simplify populating basket for a project; resilience to network issues
Basket.add() and Basket.get_distribution() were merged into a single
Basket.add_egg() as that seems to simplify the code and be easier to use
in derived classes. This was done as part of a refactoring to improve
resiliency towards network errors...

The former fill_distribution hook gives way to the finer-grained
initialize_project. The latter is expected to call add_egg to add eggs
to the basket. The base Basket keeps track of projects that have been
initialized.

As the new hook is wrapped inside a catch-all, all customization points
in use are now fail-safe, which makes both PyPIBasked and S3Basket
resilient to missing or flaky networks. This, in turn, allows us to drop
the local-retry in bootstrap's bootstrap().