Drop non-standard dependencies from S3Basket
Since botocore doesn't provide eggs in PyPI depending on it limited the
bootstrap functionality to applications hosted and depending only on
PyPI-hosted eggs.

Useless to me.

Anyway, given that we only really use two S3 APIs re-implementing (the
needed bits of) this wheel is not a big deal and gives us a much better
user story.

Finally, to reduce on roundtrips, we no longer query the bucket's
location and instead use the endpoint for whatever default region is
already set in environment variables used by other AWS tools.