Only keep track of missing entries to sys.path
When satisfying requirements, distributions explicitly added to sys.path
are now considered part of the working set. For those packages, updated
distributions will be ignored. All others will be updated from
configured sources.

This has a couple of nice properties, not the least of which is that it
solves an infinite hard-transmute loop. It also offers an easy path to
skip updates, by explicitly adding eggs to PYTHONPATH.

Resolver.reset (and thus, transmute.reset) are dropped.